requirements.txt
fastapi
python-dotenv
requests
uvicorn[standard]
numpy
tqdm
asyncio
pydantic
colorlog
tenacity
tiktoken
python-multipart
dotenv
openai
sentence-transformers
qdrant-client
pymongo
google.generativeai
cassandra-driver
prometheus-fastapi-instrumentator
prometheus_client
psutil
python-logging-loki